refactor: migrate facets handler logic to facets query builder for improved architecture
