+++ keep/api/core/facets_query_builder/get_facets_query_builder.py
@@ -3,32 +3,36 @@ from keep.api.core.cel_to_sql.sql_providers.get_cel_to_sql_provider_for_dialect 
     get_cel_to_sql_provider,
 )
 from keep.api.core.db import engine
-from keep.api.core.facets_handler.base_facets_handler import BaseFacetsHandler
-from keep.api.core.facets_handler.mysql import MySqlFacetsHandler
-from keep.api.core.facets_handler.postgresql import PostgreSqlFacetsHandler
-from keep.api.core.facets_handler.sqlite import SqliteFacetsHandler
+from keep.api.core.facets_query_builder.base_facets_query_builder import (
+    BaseFacetsQueryBuilder,
+)
+from keep.api.core.facets_query_builder.mysql import MySqlFacetsQueryBuilder
+from keep.api.core.facets_query_builder.postgresql import PostgreSqlFacetsQueryBuilder
+from keep.api.core.facets_query_builder.sqlite import SqliteFacetsHandler
 
 
-def get_facets_handler(
+def get_facets_query_builder(
     properties_metadata: PropertiesMetadata,
-) -> BaseFacetsHandler:
-    return get_facets_handler_for_dialect(engine.dialect.name, properties_metadata)
+) -> BaseFacetsQueryBuilder:
+    return get_facets_query_builder_for_dialect(
+        engine.dialect.name, properties_metadata
+    )
 
 
-def get_facets_handler_for_dialect(
+def get_facets_query_builder_for_dialect(
     dialect_name: str,
     properties_metadata: PropertiesMetadata,
-) -> BaseFacetsHandler:
+) -> BaseFacetsQueryBuilder:
     if dialect_name == "sqlite":
         return SqliteFacetsHandler(
             properties_metadata, get_cel_to_sql_provider(properties_metadata)
         )
     elif dialect_name == "mysql":
-        return MySqlFacetsHandler(
+        return MySqlFacetsQueryBuilder(
             properties_metadata, get_cel_to_sql_provider(properties_metadata)
         )
     elif dialect_name == "postgresql":
-        return PostgreSqlFacetsHandler(
+        return PostgreSqlFacetsQueryBuilder(
             properties_metadata, get_cel_to_sql_provider(properties_metadata)
         )
