+++ keep/api/core/facets_query_builder/sqlite.py
@@ -1,24 +1,16 @@
-from sqlalchemy import (
-    Column,
-    Integer,
-    String,
-    case,
-    cast,
-    func,
-    literal,
-    literal_column,
-    select,
-)
+from sqlalchemy import Integer, case, cast, func, literal, literal_column, select
 
 from keep.api.core.cel_to_sql.ast_nodes import DataType
 from keep.api.core.cel_to_sql.properties_metadata import (
     JsonFieldMapping,
     PropertyMetadataInfo,
 )
-from keep.api.core.facets_handler.base_facets_handler import BaseFacetsHandler
+from keep.api.core.facets_query_builder.base_facets_query_builder import (
+    BaseFacetsQueryBuilder,
+)
 
 
-class MySqlFacetsHandler(BaseFacetsHandler):
+class SqliteFacetsHandler(BaseFacetsQueryBuilder):
 
     def _cast_column(self, column, data_type: DataType):
         if data_type == DataType.BOOLEAN:
@@ -35,11 +27,9 @@ class MySqlFacetsHandler(BaseFacetsHandler):
         self, base_query, metadata: PropertyMetadataInfo
     ):
         column_name = metadata.field_mappings[0].map_to
-        # MySQL throws errors for JSON_TABLE if the CTE is not limited to a certain number of rows
-        base_query = base_query.limit(1_000_000).cte(f"{column_name}_base_query")
-        json_table_join = func.json_table(
-            literal_column(column_name), Column("value", String(127))
-        ).table_valued("value")
+        json_table_join = func.json_each(literal_column(column_name)).table_valued(
+            "value"
+        )
         return select(
             func.distinct(base_query.c.entity_id),
             json_table_join.c.value.label("facet_value"),
@@ -49,6 +39,6 @@ class MySqlFacetsHandler(BaseFacetsHandler):
         built_json_path = "$." + ".".join(
             [f'"{item}"' for item in field_mapping.prop_in_json]
         )
-        return func.json_unquote(
-            func.json_extract(literal_column(field_mapping.json_prop), built_json_path)
+        return func.json_extract(
+            literal_column(field_mapping.json_prop), built_json_path
         )
